Lots of changes and updates. block_splitter. sharded_storage now operates concurrently.

+++ pkg/storage/protocol/dirty_list.go
@@ -5,24 +5,24 @@ import (
 	"errors"
 )
 
-func EncodeDirtyList(blocks []uint32) []byte {
+func EncodeDirtyList(blocks []uint) []byte {
 	buff := make([]byte, 1+4+4*len(blocks))
 	buff[0] = COMMAND_DIRTY_LIST
 	binary.LittleEndian.PutUint32(buff[1:], uint32(len(blocks)))
 	for i, v := range blocks {
-		binary.LittleEndian.PutUint32(buff[(5+i*4):], v)
+		binary.LittleEndian.PutUint32(buff[(5+i*4):], uint32(v))
 	}
 	return buff
 }
 
-func DecodeDirtyList(buff []byte) ([]uint32, error) {
+func DecodeDirtyList(buff []byte) ([]uint, error) {
 	if buff == nil || len(buff) < 9 || buff[0] != COMMAND_DIRTY_LIST {
 		return nil, errors.New("Invalid packet")
 	}
 	length := binary.LittleEndian.Uint32(buff[1:])
-	blocks := make([]uint32, length)
+	blocks := make([]uint, length)
 	for i := 0; i < int(length); i++ {
-		blocks[i] = binary.LittleEndian.Uint32(buff[(5 + i*4):])
+		blocks[i] = uint(binary.LittleEndian.Uint32(buff[(5 + i*4):]))
 	}
 	return blocks, nil
 }
